pyproject.toml
[tool.poetry]
name = "weblite_framework"
version = "0.1.0"
description = ""
authors = ["Чернов Михаил <mihail.tchernov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.6"
pydantic = "2.11.7"
pydantic-settings = "2.10.1"
aioboto3 = "15.1.0"

[tool.poetry.group.test.dependencies]
pytest = "8.4.1"
pytest-asyncio = "1.1.0"
pyhamcrest = "2.1.0"

[tool.poetry.group.lint.dependencies]
isort = "6.0.1"
flake8 = "7.3.0"
flake8-quotes = "3.4.0"
flake8-annotations = "3.1.1"
flake8-docstrings = "1.7.0"
flake8-dunder-all = "0.5.0"
flake8-pyproject = "1.2.3"
black = "25.1.0"
mypy = "1.17.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "4.3.0"

[tool.pytest.ini_options]
addopts = "-ra -q"
pythonpath = "."
testpaths = "tests"
asyncio_mode = "auto"

[tool.flake8]
exclude = ["__pycache__"]
inline-quotes = "'"
max-line-length = 79
max-doc-length = 79
docstring-convention = "google"
require-annotations = 1
ignore = ["ANN101", "ANN102"]
per-file-ignores = [
    "__init__.py:D104",
     "tests/*:DALL000,D104"
]
max-complexity = 5

[tool.black]
line-length = 79
skip-string-normalization = true

[tool.isort]
include_trailing_comma = true
multi_line_output = 3
line_length = 79

[tool.mypy]
strict = true
plugins = "pydantic.mypy"
ignore_missing_imports = true
namespace_packages = true
explicit_package_bases = true

disable_error_code = [
    "import-untyped",
    "no-untyped-call"
]

[build-system]
requires = ["poetry-core==2.0.1"]
build-backend = "poetry.core.masonry.api"
